Change UberLineGraphFunc to return the new value
This is easier on bindings (returning multiple values from
a callback is not supported, and this is lighter weight
than passing a newly allocated boxed type or some such
to the callback.

I had to change the implementation to use NAN as the sentinal which
represents no data. I would have liked to continue to use -INFINITY
here, but this is not handled correctly by pygobject.